pyproject.toml
[project]
name = "proposal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.9.0",
    "openai-agents>=0.5.0",
    "tavily-python>=0.7.12",
]
